Move client pollTxImportStatus function to actions/dataset.js
